refactor(form-builder): extract edit popover in `Annotation` into a separate `AnnotationToolbarPopover` component
