+++ packages/@sanity/form-builder/src/inputs/PortableText/text/Annotation.tsx
@@ -1,4 +1,4 @@
-import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react'
+import React, {useCallback, useMemo, useRef, useState} from 'react'
 import {
   PortableTextChild,
   PortableTextEditor,
@@ -9,23 +9,11 @@ import {
 import {FOCUS_TERMINATOR} from '@sanity/util/paths'
 import {Marker, Path} from '@sanity/types'
 import styled, {css} from 'styled-components'
-import {
-  Box,
-  Button,
-  Inline,
-  Popover,
-  PopoverProps,
-  Text,
-  Theme,
-  ThemeColorToneKey,
-  Tooltip,
-  useGlobalKeyDown,
-  useTheme,
-} from '@sanity/ui'
+import {Box, Theme, ThemeColorToneKey, Tooltip} from '@sanity/ui'
 import {hues} from '@sanity/color'
-import {EditIcon, TrashIcon} from '@sanity/icons'
 import Markers from '../_legacyDefaultParts/Markers'
 import {RenderCustomMarkers} from '../types'
+import {AnnotationToolbarPopover} from './'
 
 interface AnnotationProps {
   attributes: RenderAttributes
@@ -39,8 +27,6 @@ interface AnnotationProps {
   value: PortableTextChild
   scrollElement: HTMLElement
 }
-
-const POPOVER_FALLBACK_PLACEMENTS: PopoverProps['fallbackPlacements'] = ['top', 'bottom']
 
 const Root = styled.span<{$toneKey?: ThemeColorToneKey}>(
   (props: {$toneKey?: ThemeColorToneKey; theme: Theme}) => {
@@ -76,12 +62,6 @@ const TooltipBox = styled(Box).attrs({forwardedAs: 'span'})`
   max-width: 250px;
 `
 
-const ToolbarPopover = styled(Popover)`
-  &[data-popper-reference-hidden='true'] {
-    display: none !important;
-  }
-`
-
 export function Annotation(props: AnnotationProps) {
   const {
     attributes,
@@ -99,107 +79,7 @@ export function Annotation(props: AnnotationProps) {
   const annotationRef = useRef<HTMLElement>(null)
   const editor = usePortableTextEditor()
   const markDefPath = useMemo(() => [path[0], 'markDefs', {_key: value._key}], [path, value._key])
-  const {sanity} = useTheme()
-
-  // -------------- Popover ------------- //
   const [textElement, setTextElement] = useState<HTMLSpanElement | null>(null)
-  const [open, setOpen] = useState<boolean>(false)
-  const [cursorRect, setCursorRect] = useState<DOMRect | null>(null)
-  const [selection, setSelection] = useState(null)
-  const isClosingRef = useRef<boolean>(false)
-  const rangeRef = useRef<Range | null>(null)
-
-  // This is a "virtual element" (supported by Popper.js)
-  const cursorElement = useMemo(() => {
-    if (!cursorRect) {
-      return null
-    }
-
-    return {
-      getBoundingClientRect: () => {
-        return cursorRect
-      },
-    }
-  }, [cursorRect])
-
-  useEffect(() => {
-    if (!open) {
-      return undefined
-    }
-
-    const handleScroll = () => {
-      if (rangeRef.current) {
-        setCursorRect(rangeRef.current.getBoundingClientRect())
-      }
-    }
-
-    scrollElement.addEventListener('scroll', handleScroll, {passive: true})
-
-    return () => {
-      scrollElement.removeEventListener('scroll', handleScroll)
-    }
-  }, [open, scrollElement])
-
-  // Close floating toolbar on Escape
-  useGlobalKeyDown(
-    useCallback(
-      (event) => {
-        if (event.key === 'Escape' && open) {
-          setOpen(false)
-        }
-      },
-      [open]
-    )
-  )
-
-  // Detect selection changes
-  useEffect(() => {
-    function handleSelectionChange() {
-      if (!textElement) return
-      const winSelection = window.getSelection()
-      const {anchorNode, anchorOffset, focusNode, focusOffset} = winSelection
-
-      setSelection({
-        anchorNode,
-        anchorOffset,
-        focusNode,
-        focusOffset,
-      })
-    }
-
-    document.addEventListener('selectionchange', handleSelectionChange, {passive: true})
-
-    return () => {
-      document.removeEventListener('selectionchange', handleSelectionChange)
-    }
-  }, [textElement])
-
-  // Open popover when selection is within annotations
-  useEffect(() => {
-    if (!selection) return
-    if (isClosingRef.current) return
-
-    const {anchorNode, focusNode} = selection
-
-    const annotationElement = annotationRef.current
-
-    if (annotationElement && annotationElement.contains(anchorNode) && anchorNode === focusNode) {
-      const range = window.getSelection().getRangeAt(0)
-      const rect = range.getBoundingClientRect()
-
-      rangeRef.current = range
-
-      if (rect) {
-        setCursorRect(rect)
-        setOpen(true)
-      }
-    } else {
-      setOpen(false)
-      setCursorRect(null)
-      rangeRef.current = null
-    }
-  }, [selection])
-  // -------------- End popover ------------- //
 
   const text = useMemo(
     () => (
@@ -209,8 +89,6 @@ export function Annotation(props: AnnotationProps) {
     ),
     [children]
   )
-
-  const popoverScheme = sanity.color.dark ? 'light' : 'dark'
 
   const markersToolTip = useMemo(
     () =>
@@ -232,7 +110,6 @@ export function Annotation(props: AnnotationProps) {
   )
 
   const handleEditClick = useCallback((): void => {
-    setOpen(false)
     onFocus(markDefPath.concat(FOCUS_TERMINATOR))
   }, [markDefPath, onFocus])
 
@@ -240,7 +117,6 @@ export function Annotation(props: AnnotationProps) {
     (e): void => {
       e.preventDefault()
       e.stopPropagation()
-      setOpen(false)
       PortableTextEditor.removeAnnotation(editor, type)
       PortableTextEditor.focus(editor)
     },
@@ -274,34 +150,14 @@ export function Annotation(props: AnnotationProps) {
       data-markers={markers.length > 0 ? '' : undefined}
     >
       {markersToolTip || text}
-      <ToolbarPopover
-        boundaryElement={scrollElement}
-        constrainSize
-        content={
-          <Box padding={1}>
-            <Inline space={1}>
-              <Box padding={2}>
-                <Text weight="semibold" size={1}>
-                  {type?.title || type.name}
-                </Text>
-              </Box>
-              <Button icon={EditIcon} mode="bleed" onClick={handleEditClick} padding={2} />
-              <Button
-                icon={TrashIcon}
-                mode="bleed"
-                padding={2}
-                onClick={handleRemoveClick}
-                tone="critical"
-              />
-            </Inline>
-          </Box>
-        }
-        fallbackPlacements={POPOVER_FALLBACK_PLACEMENTS}
-        open={cursorElement && open}
-        placement="top"
-        portal="editor"
-        referenceElement={cursorElement as HTMLElement}
-        scheme={popoverScheme}
+
+      <AnnotationToolbarPopover
+        textElement={textElement}
+        annotationElement={annotationRef?.current}
+        scrollElement={scrollElement}
+        onEdit={handleEditClick}
+        onDelete={handleRemoveClick}
+        title={type?.title || type.name}
       />
     </Root>
   )
